requirements.txt
PyQt6>=6.5
Pillow>=10.0
reportlab>=4.0
PyMuPDF>=1.24.0
opencv-python>=4.9.0
pyzbar>=0.1.9
qtawesome>=1.3.0
